feat: move ThreadFeedViewModel and ChannelFeedFilter+VM to commons
Move platform-agnostic versions of:
- ThreadFeedViewModel (filter was already in commons)
- ChannelFeedFilter (all types already in commons)
- ChannelFeedViewModel (wraps the now-commons filter)

Android wrappers remain in place, delegating to commons with LocalCache.

Part of the KMP iOS migration tracked in #2238.
