+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/threadview/dal/ThreadFeedViewModel.kt
@@ -22,14 +22,13 @@ package com.vitorpamplona.amethyst.ui.screen.loggedIn.threadview.dal
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import com.vitorpamplona.amethyst.commons.viewmodels.thread.ThreadFeedFilter
 import com.vitorpamplona.amethyst.model.Account
 import com.vitorpamplona.amethyst.model.LocalCache
 
 class ThreadFeedViewModel(
     account: Account,
     noteId: String,
-) : com.vitorpamplona.amethyst.commons.viewmodels.thread.LevelFeedViewModel(ThreadFeedFilter(account, noteId, LocalCache), LocalCache) {
+) : com.vitorpamplona.amethyst.commons.viewmodels.thread.ThreadFeedViewModel(account, noteId, LocalCache) {
     class Factory(
         val account: Account,
         val noteId: String,
