+++ commons/src/commonMain/kotlin/com/vitorpamplona/amethyst/commons/viewmodels/ChannelFeedViewModel.kt
@@ -18,7 +18,22 @@
  * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.vitorpamplona.amethyst.ui.screen.loggedIn.chats.publicChannels.dal
+package com.vitorpamplona.amethyst.commons.viewmodels
 
-// Re-export from commons for backwards compatibility
-typealias ChannelFeedFilter = com.vitorpamplona.amethyst.commons.ui.feeds.ChannelFeedFilter
+import com.vitorpamplona.amethyst.commons.model.Channel
+import com.vitorpamplona.amethyst.commons.model.IAccount
+import com.vitorpamplona.amethyst.commons.model.cache.ICacheProvider
+import com.vitorpamplona.amethyst.commons.ui.feeds.ChannelFeedFilter
+
+/**
+ * Platform-agnostic ChannelFeedViewModel for public channel (NIP-28) feeds.
+ *
+ * @param channel The public channel to display
+ * @param account The current user's account
+ * @param cacheProvider The cache provider for note lookups
+ */
+open class ChannelFeedViewModel(
+    val channel: Channel,
+    val account: IAccount,
+    cacheProvider: ICacheProvider,
+) : ListChangeFeedViewModel(ChannelFeedFilter(channel, account), cacheProvider)
